client/pyproject.toml
[project]
name = "wpilib-rag"
version = "0.1.0"
description = "WPILib RAG system as MCP server with version and language-specific documentation"
readme = "README.md"
authors = [
    {name = "WPILib RAG Contributors"}
]
maintainers = [
    {name = "WPILib RAG Contributors"}
]
license = {text = "MIT"}
keywords = ["wpilib", "frc", "robotics", "rag", "mcp", "documentation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "chromadb==1.3.0",
    "voyageai>=0.2.0",
    "mcp>=0.9.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
]

[project.scripts]
wpilib-rag-server = "wpilib_rag.__main__:main"
wpilib-rag-update = "wpilib_rag.database_loader:update_database_cli"

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/wpilibsuite/frc-rag-mcpserver"
Repository = "https://github.com/wpilibsuite/frc-rag-mcpserver"
Issues = "https://github.com/wpilibsuite/frc-rag-mcpserver/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build]
source-dir = "src"
